+++ src/Exception/KeyNotExistsException.php
@@ -16,7 +16,7 @@ namespace Tools\Exception;
 use Exception;
 
 /**
- * Array key not exists exception
+ * "Array key does not exist" exception
  */
 class KeyNotExistsException extends Exception
 {
@@ -24,9 +24,9 @@ class KeyNotExistsException extends Exception
      * Constructor
      * @param string $message The string of the error message
      * @param int|null $code The code of the error
-     * @param \Exception|null $previous the previous exception
+     * @param Exception|null $previous the previous exception
      */
-    public function __construct($message = 'Key does not exist', $code = null, $previous = null)
+    public function __construct($message = 'Array key does not exist', $code = null, $previous = null)
     {
         parent::__construct($message, $code, $previous);
     }