+++ src/Exception/NotReadableException.php
@@ -17,7 +17,7 @@ namespace Tools\Exception;
 use Exception;
 
 /**
- * File or directory not readable exception
+ * "File or directory is not readable" exception
  */
 class NotReadableException extends Exception
 {
@@ -25,7 +25,7 @@ class NotReadableException extends Exception
      * Constructor
      * @param string $message The string of the error message
      * @param int|null $code The code of the error
-     * @param \Exception|null $previous the previous exception
+     * @param Exception|null $previous the previous exception
      */
     public function __construct($message = 'File or directory is not readable', $code = null, $previous = null)
     {