+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # 1.x branch
 ## 1.1 branch
+### 1.1.11
+* added `IS_WIN` constant. The `is_win()` method is now deprecated and will be
+    removed in a later version;
+* added `assertIsDeprecated()` assert method;
+* all `ReflectionTrait` methods are now protected. The `setProperty()` method
+    now returns the old value or `null`;
+* `TestCaseTrait` is now deprecated and will be removed in a later version. Use
+    `TestTrait` instead. The `createSomeFiles()` method has been removed and now
+    it is a global function only for tests;
+* fixed bug for `is_url()` function with no-string values.
+
 ### 1.1.10
 * added `first_key()`, `first_value_recursive()`, `last_key()` and
     `last_value_recursive()` global functions;